Optimize the datatypes of DTOs

+++ src/main/java/lk/ijse/cmjd109/LibMgmt109/entities/LendingEntity.java
@@ -23,6 +23,6 @@ public class LendingEntity {
     private LocalDate lendingDate;
     private LocalDate returnDate;
     private Boolean isActive;
-    private Integer overDue;
+    private Long overDue;
     private Double fineAmount;
 }
